+++ src/OutputBuilderClient/Interfaces/IRebuildDetection.cs
