+++ src/OutputBuilderClient/PhotoMetadataRepository.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using ObjectModel;
+using OutputBuilderClient.Services.Emitters;
 using Scanner;
 using StorageHelpers;
